multi_lrt with categorical variables
if a linear combination of variables is continuous, multi_lrt() is fine.
But if there are only 2 or 3 unique values in the linear combination we
need to make several adjustments that ensure the split works the way we
intend it to. Some rough adjustments are introduced here. They need lots
of work.