min obs and events for splitting
obliqueRSF was a little more detailed about when to split versus
not split depending on how many obs were in a node. In addition to
having a min number of things in a leaf, obliqueRSF required a min
number of things in the node that would be split into leaves. This
makes it feasible to set lower defaults for the leaf-specific mins.