drop pkgcheck
with the ropensci review almost finished, I think it is okay to drop the
pkgcheck workflow. I would have kept it just cause I like having lots of
badges on the readme, but there is an issue with using Rmd files for
examples that makes aorsf fail the pkgcheck test every time.