the moby dick of bugs has been caught
for a long time I was not sure why I would get very slightly different
results if I predicted row by row in a dataset versus the whole thing.
It turned out to be an issue in the predict_value routine where the case
of the final value in pred_leaf being a unique leaf would lead to that
value getting the leaf data from the prior leaf. I had to introduce
several new features into the predict function to catch this error and
those features now need to be ironed out.