soften requirements on binary y variables
if a logical value is provided for an outcome, coerce it to an integer
so that it can safely be passed to C++ without breaking expectations.
This is easy enough for outcomes because outcome types are only checked
prior to growing a forest. It would be more difficult to implement this
safely for predictors.